+++ R/aggregate.R
@@ -1,23 +1,26 @@
 #---- Aggregate an index ----
 aggregate.ind <- function(x, pias, na.rm = FALSE, r = 1, ...) {
   if (!is_pias(pias)) {
-    stop(gettext("'pias' must be a price index aggregation structure; use aggregation_structure() to make one"))
+    stop("'pias' must be a price index aggregation structure; use ",
+         "aggregation_structure() to make one")
   }
   # helpful functions
   price_update <- factor_weights(r)
   gen_mean <- generalized_mean(r)
   aw <- transmute_weights(r, 1)
   # functions to aggregate index values and contributions
-  # 'i' is defined in the loop below; it's used to loop over the height of 'pias'
-  aggregate_index <- function(z) {
+  # 'i' is defined in the loop below; it's used to loop over the height of
+  # 'pias'
+  aggregate_index <- function(z, i) {
     gen_mean(rel[[i - 1L]][z], w[[i - 1L]][z], na.rm = na.rm)
   }
   aggregate_contrib <- if (x$has_contrib) {
-    function(z) {
-      unlist(Map("*", con[[i - 1L]][z], scale_weights(aw(rel[[i - 1L]][z], w[[i - 1L]][z]))))
+    function(z, i, con) {
+      unlist(Map("*", con[[i - 1L]][z],
+                 scale_weights(aw(rel[[i - 1L]][z], w[[i - 1L]][z]))))
     }
   } else {
-    function(z) numeric(0L)
+    function(z, i, con) numeric(0L)
   }
   # put the aggregation weights upside down to line up with pias
   w <- rev(weights(pias, na.rm = na.rm))
@@ -36,8 +39,8 @@ aggregate.ind <- function(x, pias, na.rm = FALSE, r = 1, ...) {
     }
     # loop over each level in the pias from the bottom up and aggregate
     for (i in seq_along(rel)[-1L]) {
-      rel[[i]] <- vapply(pias$child[[i - 1L]], aggregate_index, numeric(1L))
-      con[[i]] <- lapply(pias$child[[i - 1L]], aggregate_contrib)
+      rel[[i]] <- vapply(pias$child[[i - 1L]], aggregate_index, numeric(1L), i)
+      con[[i]] <- lapply(pias$child[[i - 1L]], aggregate_contrib, i, con)
     }
     # parental imputation
     if (na.rm) {
@@ -52,11 +55,11 @@ aggregate.ind <- function(x, pias, na.rm = FALSE, r = 1, ...) {
     x$contrib[[t]] <- unlist(rev(con), recursive = FALSE)
     # price update weights for all periods after the first
     if (x$chainable) {
-      pias$weights <- price_update(index[pias$eas], w[[1L]]) 
+      pias$weights <- price_update(index[pias$eas], w[[1L]])
     }
   }
   x$levels <- pias$levels
-  x$r <- r
+  x$r <- as.numeric(r)
   x$pias <- pias[c("child", "parent", "eas", "height")]
   structure(x, class = c("agg_ind", "ind"))
 }
@@ -73,8 +76,8 @@ vcov.agg_ind <- function(object, repweights, mse = TRUE, ...) {
   r <- object$r
   # template aggregation structure with no weights for each bootstrap replicate
   pias <- aggregate2pias(object, numeric(length(eas)))
-  # matrix aggregation is much faster than aggregate.ind(), but needs to be done 
-  # with a chained index
+  # matrix aggregation is much faster than aggregate.ind(), but needs to be
+  # done with a chained index
   elem <- as.matrix(chain(object[eas, ]))
   repindex <- lapply(seq_len(n), function(i) {
     pias$weights[] <- repweights[, i]
@@ -85,8 +88,9 @@ vcov.agg_ind <- function(object, repweights, mse = TRUE, ...) {
   })
   # it's easier to calculate the variance with an array of indexes
   dimnm <- list(upper, object$time, seq_len(n))
-  repindex <- array(unlist(repindex, use.names = FALSE), dim = lengths(dimnm), dimnames = dimnm)
-  # mse = TRUE is the default for variance estimation in SAS, 
+  repindex <- array(unlist(repindex, use.names = FALSE),
+                    dim = lengths(dimnm), dimnames = dimnm)
+  # mse = TRUE is the default for variance estimation in SAS,
   # but not the survey package
   centre <- if (mse) {
     as.matrix(object)[upper, , drop = FALSE]